refactor: 구매 섹션을 개요~원두 라인업 사이로 이동, 원두 클릭으로 가격 전환
- RoasteryBuyAndBeans 컴포넌트 신규 생성 (PurchaseSection + BeanList 통합)
- 원두 라인업 아이템 클릭 시 구매 채널 가격 실시간 전환
- RoasteryDetail에서 RoasteryPurchaseFlow 제거, RoasteryBuyAndBeans로 교체
- BeanTabs, RoasteryPurchaseFlow 삭제 (대체됨)
